# docs/openapi.yaml
...
openapi: 3.1.1
info:
  title: AR.IO Gateway
  version: 0.0.1
  x-logo:
    url: 'https://ar.io/wp-content/uploads/2022/02/ar-io-FINAL-LOGO-SHADOW.png'
  description: |
    AR.IO Gateways are the interface between users and the Permaweb. Each Gateway acts like a “Permaweb Service
    Provider” and supports multiple, value-added, low-trust services for users and applications.

    The AR.IO HTTP API is served by all AR.IO Gateways, and is used by all Arweave clients looking to read, write and query data.  It supports the following capabilities
    - seeding layer one transactions to the greater Arweave network
    - caching and serving chain and transaction data
    - resolving friendly names to transaction data with the Arweave Name System
    - indexing and querying layer one and two transactions
    - proxy for Arweave nodes serving the Arweave HTTP API

    ## Errors

    The API uses standard HTTP status codes to indicate the success or failure of the API call.

    ## Authentication

    All inbound transactions to the Arweave network must contain a valid signature from an Arweave wallet with appropriate amount of AR tokens.  At this time, clients are not required to sign their requests for outbound data from the Arweave network.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: apiToken
      description: ADMIN_API_KEY set in your .env file.
  schemas:
...
    Base64Url43:
      type: string
      pattern: '^[0-9a-zA-Z_-]{43}$'
...
    Chunk:
      type: object
      properties:
        tx_path:
          $ref: '#/components/schemas/Base64Url'
        packing:
          type: string
          # TODO add all types
          pattern: '^unpacked$'
        data_path:
          $ref: '#/components/schemas/Base64Url'
        chunk:
          $ref: '#/components/schemas/Base64Url'
      required:
        - tx_path
        - packing
        - data_path
        - chunk
...
paths:
...
  '/chunk/{offset}':
    get:
      tags: [Chunks]
      summary: Get chunk offset information.
      description: Fetches information about the size and offset of a specified chunk.
      parameters:
        - name: offset
          in: path
          required: true
          schema:
            type: integer
        - name: If-None-Match
          in: header
          required: false
          description: ETag value for conditional request. Returns 304 if chunk hash matches.
          schema:
            type: string
      responses:
        '200':
          description: |-
            200 response
          headers:
            ETag:
              description: Entity tag based on chunk hash (when available)
              schema:
                type: string
            Content-Digest:
              description: RFC 9530 compliant digest header with SHA-256 (when available)
              schema:
                type: string
                example: 'sha-256=:YjNkZjM1YThhNjIxOGY2MTZlNTQ3YzFmZGE2YzE4ZjE=:'
            Content-Length:
              description: Size of the response body in bytes
              schema:
                type: integer
            X-Cache:
              description: |-
                Cache status indicating whether chunk was served from local cache (HIT) or retrieved from external source (MISS).
                Always present.
              schema:
                type: string
                enum: [HIT, MISS]
              example: HIT
            X-AR-IO-Chunk-Source-Type:
              description: Source type of the chunk data
              schema:
                type: string
            X-AR-IO-Chunk-Host:
              description: Host that provided the chunk
              schema:
                type: string
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/Chunk'
        '304':
          description: Not Modified - chunk hash matches If-None-Match header
        '400':
          description: Invalid offset parameter
        '404':
          description: Chunk not found
    head:
      tags: [Chunks]
      summary: Get chunk metadata without body.
      description: Returns the same headers as GET but without the response body. Useful for checking chunk existence and metadata.
      parameters:
        - name: offset
          in: path
          required: true
          schema:
            type: integer
        - name: If-None-Match
          in: header
          required: false
          description: ETag value for conditional request. Returns 304 if chunk hash matches.
          schema:
            type: string
      responses:
        '200':
          description: |-
            200 response with headers only
          headers:
            Content-Type:
              description: Media type of the response
              schema:
                type: string
            ETag:
              description: Entity tag based on chunk hash (when available)
              schema:
                type: string
            Content-Digest:
              description: RFC 9530 compliant digest header with SHA-256 (when available)
              schema:
                type: string
                example: 'sha-256=:YjNkZjM1YThhNjIxOGY2MTZlNTQ3YzFmZGE2YzE4ZjE=:'
            Content-Length:
              description: Size of the response body in bytes
              schema:
                type: integer
            X-Cache:
              description: |-
                Cache status indicating whether chunk was served from local cache (HIT) or retrieved from external source (MISS).
                Always present.
              schema:
                type: string
                enum: [HIT, MISS]
              example: HIT
            X-AR-IO-Chunk-Source-Type:
              description: Source type of the chunk data
              schema:
                type: string
            X-AR-IO-Chunk-Host:
              description: Host that provided the chunk
              schema:
                type: string
        '304':
          description: Not Modified - chunk hash matches If-None-Match header
        '400':
          description: Invalid offset parameter
        '404':
          description: Chunk not found
...
  '/graphql':
    post:
      tags: [Index Querying]
      summary: Query indexed data using GraphQL
      description: |
        GraphQL endpoint for querying indexed transaction and block data. Supports:
        - Transaction queries by ID, owner, recipient, tags, and bundle
        - Block queries by ID and height
        - Pagination and sorting
        - Rich metadata including sizes, content types, and signatures

        See the GraphQL Playground at `/graphql` for full schema documentation and interactive querying.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: GraphQL query string
                variables:
                  type: object
                  description: Query variables
              example:
                query: |
                  query {
                    transaction(
                      id: "D6I-Ke9oLg7sU_PS4oADQmnyFRa1jj2Zs1DP9BltrFs"
                    ) {
                      signature
                      owner {
                        address
                      }
                      bundledIn {
                        id
                      }
                      block {
                        height
                      }
                    }
                  }
      responses:
        '200':
          description: Successful GraphQL response
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    description: Query results
                    example: |
                      {
                        "transaction": {
                          "signature": "4wh6PDJH6eyCKAWZoW26T0qt5NNjnVgyPjLs5HDR7fYgOyzEZF8FyIJiz01IRVvmXC9WZKvy2uLGiIdW4GtUNhs",
                          "owner": {
                            "address": "WbE3R_GcAalEP-VonRpL2P5TG30PJRA1-Tj20E_xIK0"
                          },
                          "bundledIn": {
                            "id": "yU5ilscWlDecDhUmNcd24zqe9HmNZ31h31BxC_VcXIM"
                          },
                          "block": null
                        }
                      }
                  errors:
                    type: array
                    description: Query errors if any
                    items:
                      type: object
                      properties:
                        message:
                          type: string
                        locations:
                          type: array
                          items:
                            type: object
                            properties:
                              line:
                                type: integer
                              column:
                                type: integer
                        path:
                          type: array
                          items:
                            type: string
        '400':
          description: Invalid GraphQL query
        '500':
          description: Server error

  # ArNS
  '/ar-io/resolver/{name}':
    get:
      tags: [ArNS]
      summary: Get ArNS name resolution.
      description: Get detailed information of a specified ArNS name.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful resolution of ArNS name
          headers:
            x-arns-resolved-id:
              schema:
                $ref: '#/components/schemas/Base64Url43'
              description: The resolved transaction ID
            x-arns-ttl-seconds:
              schema:
                type: string
              description: Time-to-live in seconds for this resolution
            x-arns-process-id:
              schema:
                type: string
              description: AO process ID
            x-arns-resolved-at:
              schema:
                type: string
              description: Unix timestamp when the name was resolved
            x-arns-record-index:
              schema:
                type: number
              description: Index of the name in the sorted list of names
            x-arns-undername-limit:
              schema:
                type: number
              description: The undername limit for the base ArNS name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArNSResolution'
        '404':
          description: ArNS name not found.

  # Admin
  '/ar-io/admin/debug':
    get:
      tags: [Admin]
      summary: Admin debugging information.
      description: Get detailed information about the current operational state of your AR.IO Gateway, including information about any current warnings or errors.
      operationId: adminDebug
      responses:
        '200':
          description: Successful operation.
        '401':
          description: Unauthorized, API key likely incorrect.
      security:
        - bearerAuth: []
  '/ar-io/admin/queue-tx':
    post:
      tags: [Admin]
      summary: Prioritize a transaction or bundle in the queue.
      description: Stage a specific TX ID as priority for your Gateway to locate and index. This will trigger and queue bundle processing if the transaction is a bundle, and your Gateway is configured to unbundle and index.
      operationId: adminQueueTx
      requestBody:
        $ref: '#/components/requestBodies/QueueTxRequest'
      responses:
        '200':
          $ref: '#/components/responses/QueueTxSuccess'
        '400':
          $ref: '#/components/responses/MissingId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        - bearerAuth: []
  '/ar-io/admin/queue-bundle':
    post:
      tags: [Admin]
      summary: Queue a bundle for indexing, bypassing any filter settings by default.
      description: Queue a bundle for indexing, bypassing any filter settings by default.
      operationId: adminQueueBundle
      requestBody:
        $ref: '#/components/requestBodies/QueueTxRequest'
      responses:
        '200':
          $ref: '#/components/responses/QueueTxSuccess'
        '400':
          $ref: '#/components/responses/MissingIdOrBypassFilter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        - bearerAuth: []
  '/ar-io/admin/queue-data-item':
    post:
      tags: [Admin]
      summary: Queue data items for indexing
      description: |
        Queue one or more data items for indexing. This endpoint
        accepts an array of data item headers and prioritizes them for processing.
      operationId: adminQueueDataItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  data_size:
                    type: integer
                    description: Size of the data item in bytes
                  id:
                    $ref: '#/components/schemas/Base64Url43'
                    description: Data item ID
                  owner:
                    $ref: '#/components/schemas/Owner'
                    Description: Data item signer's public key
                  owner_address:
                    $ref: '#/components/schemas/Base64Url43'
                    description: Normalized address
                  signature:
                    $ref: '#/components/schemas/Signature'
                    description: Data item signature
                  tags:
                    $ref: '#/components/schemas/TagList'
                    description: Optional tags associated with the data item
                  content_type:
                    type: string
                    description: Optional content type of the data item
                  target:
                    $ref: '#/components/schemas/Base64Url43'
                    description: Optional target address
                  anchor:
                    $ref: '#/components/schemas/Base64Url43'
                    description: Optional anchor
                required:
                  - data_size
                  - id
                  - owner
                  - owner_address
                  - signature
              minItems: 1
      responses:
        '200':
          description: Data items successfully queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Data item(s) queued'
        '400':
          description: Invalid request body
          content:
            text/plain:
              schema:
                type: string
                example: 'Must provide array of data item headers'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  '/ar-io/admin/block-data':
    put:
      tags: [Admin]
      summary: Blocks transactions or data-items so your AR.IO Gateway will not serve them.
      description: |
        Submits a TX ID/data-item ID or sha-256 content hash for content you do not want your AR.IO Gateway to serve. Once submitted, your Gateway will not respond to requests for these transactions or data-items.


        WARNING - Testing a TX ID here WILL result in that data being blocked by your Gateway.
      operationId: adminBlockData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: TX ID for a transaction you want to block.
                notes:
                  type: string
                  description: Any notes or comments related to the block data. Documentation purposes only.
                source:
                  type: string
                  description: The source from which the block data was obtained. Documentation purposes only.
            examples:
              example1:
                value:
                  id: '3lyxgbgEvqNSvJrTX2J7CfRychUD5KClFhhVLyTPNCQ'
                  notes: 'This content is offensive'
                  source: 'Public Block list'
      responses:
        '200':
          description: Successful operation.
        '401':
          description: Unauthorized, API key likely incorrect.
      security:
        - bearerAuth: []
...
  '/ar-io/admin/unblock-name':
    put:
      tags: [Admin]
      summary: Unblock an ARNS name.
      description: Unblock an ARNS name so your AR.IO Gateway will serve it again.
      operationId: adminUnblockName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name you want to unblock.
      responses:
        '200':
          description: Successful operation.
        '401':
          description: Unauthorized, API key likely incorrect.
      security:
        - bearerAuth: []
...